dropdown and definition
