+++ packages/perseus/src/widget-ai-utils/dropdown/dropdown-ai-utils.test.ts
@@ -3,7 +3,10 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./dropdown-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
+import type {PerseusDropdownUserInput} from "../../validation.types";
 import type {UserEvent} from "@testing-library/user-event";
 
 export const question1: PerseusRenderer = {
@@ -38,11 +41,34 @@ export const question1: PerseusRenderer = {
     },
 };
 
-describe("dropdown widget", () => {
+describe("Dropdown AI utils", () => {
     let userEvent: UserEvent;
     beforeEach(() => {
         userEvent = userEventLib.setup({
             advanceTimers: jest.advanceTimersByTime,
+        });
+    });
+
+    it("it returns JSON with the expected format and fields", () => {
+        const renderProps: any = {
+            choices: ["Pickles", "Tomato", "Onion", "Lettuce"],
+        };
+
+        const userInput: PerseusDropdownUserInput = {
+            value: 3,
+        };
+
+        const resultJSON = getPromptJSON(renderProps, userInput);
+
+        expect(resultJSON).toEqual({
+            type: "dropdown",
+            options: {
+                items: ["Pickles", "Tomato", "Onion", "Lettuce"],
+            },
+            userInput: {
+                // Offset to account for placeholder
+                selectedIndex: 2,
+            },
         });
     });
